+++ Sources/Networking/OfferingsAPI.swift
@@ -15,21 +15,16 @@ import Foundation
 
 class OfferingsAPI {
 
-    typealias ResponseHandler<Response> = (Result<Response, BackendError>) -> Void
-
     typealias IntroEligibilityResponseHandler = ([String: IntroEligibility], BackendError?) -> Void
-    typealias OfferSigningResponseHandler = ResponseHandler<PostOfferForSigningOperation.SigningData>
-    typealias OfferingsResponseHandler = ResponseHandler<OfferingsResponse>
-    typealias ProductEntitlementMappingResponseHandler = ResponseHandler<ProductEntitlementMappingResponse>
+    typealias OfferSigningResponseHandler = Backend.ResponseHandler<PostOfferForSigningOperation.SigningData>
+    typealias OfferingsResponseHandler = Backend.ResponseHandler<OfferingsResponse>
 
     private let offeringsCallbacksCache: CallbackCache<OfferingsCallback>
-    private let productEntitlementMappingCallbacksCache: CallbackCache<ProductEntitlementMappingCallback>
     private let backendConfig: BackendConfiguration
 
     init(backendConfig: BackendConfiguration) {
         self.backendConfig = backendConfig
         self.offeringsCallbacksCache = .init()
-        self.productEntitlementMappingCallbacksCache = .init()
     }
 
     func getOfferings(appUserID: String,
@@ -83,23 +78,6 @@ class OfferingsAPI {
                                                                         postOfferForSigningData: postOfferData,
                                                                         responseHandler: completion)
         self.backendConfig.operationQueue.addOperation(postOfferForSigningOperation)
-    }
-
-    func getProductEntitlementMapping(withRandomDelay randomDelay: Bool,
-                                      completion: @escaping ProductEntitlementMappingResponseHandler) {
-        let factory = GetProductEntitlementMappingOperation.createFactory(
-            configuration: self.backendConfig,
-            callbackCache: self.productEntitlementMappingCallbacksCache
-        )
-
-        let callback = ProductEntitlementMappingCallback(cacheKey: factory.cacheKey, completion: completion)
-        let cacheStatus = self.productEntitlementMappingCallbacksCache.add(callback)
-
-        self.backendConfig.addCacheableOperation(
-            with: factory,
-            withRandomDelay: randomDelay,
-            cacheStatus: cacheStatus
-        )
     }
 
 }
